+++ src/js/render_challenges.js
@@ -1,0 +1,6 @@
+export default async function renderChallenges(loadData, filterFunction, createCardFunction) {
+    let data = await loadData();
+    data = filterFunction(data);
+    createCardFunction(data);
+    return data;
+  }
